Extract tests for isSyntaxError to own file
