+++ lib/referee.test.js
@@ -4,43 +4,6 @@ var referee = require("../lib/referee");
 var testHelper = require("./test-helper");
 var sinon = require("sinon");
 
-testHelper.assertionTests("assert", "isSyntaxError", function(
-    pass,
-    fail,
-    msg,
-    error
-) {
-    function captureArgs() {
-        return arguments;
-    }
-
-    fail("for Error", new Error("not pie"));
-    pass("for SyntaxError", new SyntaxError("apple pie"));
-    fail("for string", "apple pie");
-    fail("for array", []);
-    fail("for object", {});
-    fail("for arguments", captureArgs());
-    msg(
-        "fail with descriptive message",
-        "[assert.isSyntaxError] Expected {  } to be a SyntaxError",
-        {}
-    );
-    msg(
-        "fail with custom message",
-        "[assert.isSyntaxError] Nope: Expected {  } to be a SyntaxError",
-        {},
-        "Nope"
-    );
-    error(
-        "for object",
-        {
-            code: "ERR_ASSERTION",
-            operator: "assert.isSyntaxError"
-        },
-        {}
-    );
-});
-
 testHelper.assertionTests("assert", "isTypeError", function(
     pass,
     fail,
